Implementação do domínio de categoria

+++ server_v2/src/Api.Domain/Models/BaseModel.cs
@@ -2,12 +2,21 @@
 
 namespace Domain.Models
 {
+    /// <summary>
+    /// Objeto de modelo base.
+    /// </summary>
     public class BaseModel
     {
-        public long Id { get; set; }
+        /// <summary>
+        /// Identificador do registro.
+        /// </summary>
+        public int Id { get; set; }
 
         private DateTime _dataCriacao;
 
+        /// <summary>
+        /// Data de criação do registro.
+        /// </summary>
         public DateTime DataCriacao
         {
             get { return _dataCriacao; }
@@ -19,6 +28,9 @@ namespace Domain.Models
 
         private DateTime _dataAlteracao;
 
+        /// <summary>
+        /// Data de atualização do registro.
+        /// </summary>
         public DateTime DataAlteracao
         {
             get { return _dataAlteracao; }
